sync

+++ homogeneous_nucleation/manage_job.py
@@ -113,7 +113,10 @@ def gather() -> None:
             src_path = path_src_dir / src_filename
             dst_path = path_dst_dir / dst_filename
             dst_path.parent.mkdir(exist_ok=True, parents=True)
-            shutil.copyfile(src_path, dst_path)
+            if src_path.exists():
+                shutil.copyfile(src_path, dst_path)
+            else:
+                raise RuntimeError(f"{src_path} does not exist.")
 
     command = "tar cvf result.tar result"
     subprocess.run(shlex.split(command))
